chore: Move string matcher to utils
